Deleted update comp from admin dash
